+++ tests/integration/test_pw.py
@@ -318,6 +318,16 @@ def test_shell(tmp_project):
     assert not proc_result.stderr.decode("utf-8")
 
 
+def test_venv_dir(tmp_project):
+    project_dir, env = tmp_project
+    assert Path(project_dir, f"{SCRIPT_PREFIX}pw").is_file()
+
+    cmd = f"{SCRIPT_PREFIX}pw -q venv-pycowsay from-venv-dir"
+    proc_result = subprocess.run(cmd, shell=True, capture_output=True, cwd=project_dir, env=env, check=False)
+    assert "from-venv-dir" in proc_result.stdout.decode("utf-8")
+    assert Path(project_dir, ".venv").is_dir()
+
+
 def load_toml(path):
     with path.open() as f:
         return tomlkit.load(f)